fucking veel gedaan, nu nog views maken

+++ Project/Logic/Logic/LandCollectie.cs
@@ -13,7 +13,7 @@ namespace Logic
             landen = new List<Land>();
         }
 
-        public void VoegLandToe(string naam, int inwoners, double sb, double db)
+        public void VoegLandToe(string naam, int inwoners, decimal sb, decimal db)
         {
             Land land = new Land(naam, inwoners, sb, db);
             this.landen.Add(land);
