+++ Project/Logic/Logic/MaatregelCategorie.cs
@@ -14,7 +14,7 @@ namespace Logic
             maatregels = new List<Maatregel>();
         }
 
-        public void MaatregelToevoegen(string naam, double sbf, double dbf, int ernst, double bg, double gbg, double sg)
+        public void MaatregelToevoegen(string naam, decimal sbf, decimal dbf, int ernst, decimal bg, decimal gbg, decimal sg)
         {
             Maatregel maatregel = new Maatregel(naam, sbf, dbf, ernst, bg, gbg, sg);
             this.maatregels.Add(maatregel);
